Extract dry-run platform logic into dedicated file
- Dry-run methods (pullDeckhousePlatformDryRun, extractImageDigestsFromRemote) live in platform_dryrun.go for clear separation from the normal pull path
- Grouped dry-run output with per-category headers and sorted image refs
